Extract decryptPhoneNumber to its own file
