+++ api-js/src/validators/__tests__/decrypt-phone-number.test.js
@@ -1,0 +1,1 @@
+import { decryptPhoneNumber } from '../decrypt-phone-number'
