+++ api-js/src/validators/index.js
@@ -1,2 +1,3 @@
 export * from './cleanse-input'
+export * from './decrypt-phone-number'
 export * from './slugify'
